chore: move RangeCheckBuiltinRunner to own file
